core/standards/css3/image-fit/js/inherited/001.html

<!doctype html>
<title>inherited object-fit</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div style='-o-object-fit:contain'><div></div></div>
<script>
test(function() {
  var e = document.body.firstChild.firstChild;
  assert_equals(getComputedStyle(e, '').OObjectFit, 'auto');
  assert_equals(getComputedStyle(e, '').OObjectPosition, '50% 50%');
});
</script>
